(#10) 升级不成功，应用程序状态变为“未安装” 升级不成功时，将当前应用的本地存储文件删除，然后添加重新写入旧版本

+++ js/app.js
@@ -19,7 +19,6 @@ function get_local_app_list(func, error_func)
 {
   console.log("get_local_app_list: ");
   cmd = " cd " + SYS_DATA_DIR + "; cat * ";
-
   get_local_service(cmd, func, error_func);
 }
 
@@ -114,7 +113,6 @@ function app_get_status(id, func) {
     status = "not-defined";
     func(status);
   }
-
   get_local_service(cmd, callback, error_func);
 }
 
@@ -200,11 +198,11 @@ function app_button_change_status($btn, id, status)
 
   $btn.off('click').click(function () {
      if (status == "not-installed")
-       app_install($btn, id);
+       app_install($btn, id,"new_install");
      else if (status == "installed")
       /* 不可操作 */;
      else if (status == "need-updated")
-       app_update($btn, id);
+       app_update($btn, id,"update");
      else if (status == "downloading")
        /* 不可操作 */;
      else if (status == "checking-download-file")
@@ -225,8 +223,7 @@ function app_button_change_status($btn, id, status)
    {
      app_button_change_status($btn, id, "not-installed");
      warning_message(app_get_name(id) + app_get_button_text(status));
-
-     app_clear_status(id);
+     app_clear_status($btn,id);
    }
 }
 
@@ -268,57 +265,67 @@ function app_inc_download_count(id)
 /*
  * 安装一个应用
  */
-function app_install($btn, id)
+function app_install($btn, id,install_type)
 {
-  $btn.css("disabled", "true");
-  console.log("app_install: " + id);
+    /*获取老的版本*/
+   var old_verson_cmd = "cut -d: -f2 /opt/app/db/"+id;
+   var old_version_callback = function(data, errno){
+    if (errno == 0){
+    install_type = install_type + "-" + data;
+    };
+    $btn.css("disabled", "true");
+    console.log("app_install: " + id);
 
-  app_button_change_status($btn, id, "installing");
+    app_button_change_status($btn, id, "installing");
 
-  var version= app_get_server_version(id);
-  var download_url = app_get_download_url(id);
-  var download_file = app_get_download_local_file(id);
-  var md5 = app_get_md5(id);
+    var version= app_get_server_version(id);
+    var download_url = app_get_download_url(id);
+    var download_file = app_get_download_local_file(id);
+    var md5 = app_get_md5(id);
 
-  /* 处理“[FILE]”通配符 */
-  var install_script = app_get_install_script(id)
+    /* 处理“[FILE]”通配符 */
+    var install_script = app_get_install_script(id)
               .replace("[FILE]", download_file);
 
-  cmd = INSTALL_SCRIPT + " "
-             + id + " "
-             + version + " "
-             + download_url + " "
-             + download_file + " "
-             + md5
-             + " \"" + install_script + "\" ";
-  console.log(cmd);
+    cmd = INSTALL_SCRIPT + " "
+              + id + " "
+              + version + " "
+              + download_url + " "
+              + download_file + " "
+              + md5
+              + " \"" + install_script + "\" "
+              + install_type;
+    console.log(cmd);
 
-  var callback = function(data, errno) {
-    if (errno != 0)
-    {
-      // 弹出错误提示，自动消失
-      console.log("安装应用程序不正常（返回值为" + data[0] + "）！");
+    var callback = function(data, errno) {
+      if (errno != 0)
+      {
+        // 弹出错误提示，自动消失
+        console.log("安装应用程序不正常（返回值为" + data[0] + "）！");
+      }
+      else
+      {
+        success_message(app_get_name(id) + "安装成功");
+        app_inc_download_count(id);
+      }
     }
-    else
-    {
-      success_message(app_get_name(id) + "安装成功");
-      app_inc_download_count(id);
-    }
+
+    console.log("开始安装");
+    get_local_service(cmd, callback);
+
+    /* 轮询更新按钮状态 */
+    poll_status($btn, id);
   }
+    get_local_service(old_verson_cmd,old_version_callback);
 
-  console.log("开始安装");
-  get_local_service(cmd, callback);
-
-  /* 轮询更新按钮状态 */
-  poll_status($btn, id);
 }
 
 /*
  * 升级一个应用
  */
-function app_update($btn, id)
+function app_update($btn, id, install_type)
 {
-  app_install($btn, id);
+    app_install($btn, id,install_type);
 }
 
 /*
@@ -368,16 +375,34 @@ function initButton($btn, id) {
   refresh_app_status($btn, id);
 }
 
-function app_clear_status(id)
+function app_clear_status($btn,id)
 {
-  console.log("clear app: " + id);
-  cmd = " rm -f " + SYS_DATA_DIR + id;
+  /*检测app目前状态：属于安装还是更新？*/
+  console.log("获取app本机存储文件内容："+id);
+  var check_cmd = "cat /opt/app/db/"+id;
+  var check_callback = function(data,errno){
+    var data_array = data.split(":");
+    var old_version = data_array[3].split("-");
+    var cmd = "";
 
-  var callback = function(data, errno) {
-    ;
-  }
+    /*根据检测结果判断，是否删除本机记录*/
+    if(data_array[3] == "new_install"){
+      console.log("clear app: " + id);
+      cmd = " rm -f " + SYS_DATA_DIR + id;
+     }
+     if(old_version[0] == "update"){
+      cmd = " rm -f " + SYS_DATA_DIR + id + ";" + "echo " + id +":" + old_version[1] + ":installed:new_install  >"+ SYS_DATA_DIR + id;
+     }
+      var callback = function(data, errno) {
+      refresh_app_status($btn,id);
+      };
 
-  get_local_service(cmd, callback);
+      get_local_service(cmd, callback);
+
+  };
+
+  get_local_service(check_cmd, check_callback);
+
 }
 
 function refresh_app_status($btn, id) {
